example/prog-guide/rb1d: Improvements to RB1D
Make a number of improvements to the RB1D example:

1. Ensure the pseudo-random init of the data array (and hence the
   convergeance behavior) is independent of job size.
2. Adjust the neighbor fetch declarations for clarity
3. Remove an unnecessary barrier
4. Rename the `start` variable to `phase`
5. Make array size 64-bit clean

This will require some minor corresponding changes to PG prose

+++ example/prog-guide/rb1d-check.cpp
@@ -1,9 +1,9 @@
 //SNIPPET
-bool check_convergence(double *u, int n_local, const double EXPECTED_VAL,
+bool check_convergence(double *u, long n_local, const double EXPECTED_VAL,
                        const double EPSILON, long stepi)
 {
   double err = 0;
-  for (int i = 1; i < n_local - 1; i++)
+  for (long i = 1; i < n_local - 1; i++)
     err = max(err, fabs(EXPECTED_VAL - u[i]));
   // upcxx collective to get max error over all processes
   double max_err = upcxx::reduce_all(err, upcxx::op_fast_max).wait();
